requirements.txt
--pre
py-cord==2.0.0rc1
python-dotenv==0.20.0
sentry-sdk>=1.5